PRD: Allow clearing FIR bits on Unit Checkstops
There is no planned support for unit checkstops in P9, so we don't
have any code in place for properly disabling checkstopped units.
It shouldn't matter, but when we get a bit mistakenly configured
as a Unit CS, we end up with a flood of errors for the uncleared
bit.

Change-Id: Ia7104b31c988434ca6df74889afad14c7ecd158b
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/42613
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Caleb N. Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Zane C. Shelley <zshelle@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/42635
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>