Move disabling PSI CEC interrupts to the common path for INT reset
In DD1 we used a SW reset to reset the INT unit, in DD2 we will use
the HW resets. In both paths we need to disable the PSI CEC intrs.
This commit moves the code that disables PSI CEC interrupts to before
the code paths diverge in the resetIntUnit func in intrrp.C

Change-Id: I2ebdbf374e7a2561151a138c6304ef31da1af4aa
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/42850
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>