Clear out OCC Flag register during pm_reset
Following a powman_suspend we were seeing issues where the OCC was
resetting itself up in runtime because the pgpe still had the flags
set to put itself in suspend. So when PHYP started the GPEs PGPE went
into suspend and caused the OCC to reset into Safe mode. Clearing
this register will avoid the pgpe going into suspend.

Change-Id: Id0c64007ccfbbb8441f28fef543f6d232c06ffc1
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/43399
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Reviewed-by: Michael S. Floyd <mfloyd@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/43404
Reviewed-by: Hostboot Team <hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>