Enable SBEIO PSU calls to Open/Close Unsecure Memory Regions
This commit enables the SBEIO PSU calls to the SBE to open and
close Unsecure Memory Regions now that the SBE team supports that
API.  Their support, however, does not actually whitelist/blacklist
the memory at this time. This commit also gracefully handles the
errors from the SBE saying this support is not there.

Change-Id: Ia9367b17e403bf759d778de5f4bc7fca3544397b
RTC:174970
CMVC-Prereq: 1030098
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/44018
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Stephen M. Cprek <smcprek@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>