Disable SBE start recovery code
The changes that just went in to enable recovery from non-fatal
SBE issues have had an unintended consequence of making live
debug of SBE failures much more difficult.  This change will
temporarily disable the code path until we put in a real
solution that works for live and automated scenarios.

Change-Id: I17b06ec8aaa8f593af2671bccb55554380a28ea8
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/44026
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Reviewed-by: Elizabeth K. Liner <eliner@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>