p9_sbe_lpc_init: Fix LPC bus LRESET for DD2
Change the order of operations so that first the LPC Master is reset,
then the bus reset is driven with a proper delay. Previously the bus reset
duration depended on the code path taken (DD1 vs. DD2) and was likely too
short on DD2.

Also extract the various parts of the sequence into individual functions to
improve readability.

Change-Id: I55cd750735ae21131c57c185ffc8a4fa1392b2dc
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/44416
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>