Keep masterVersionCompare() call out of OpenPOWER SBE Update re-IPL path
The call to sbeDoReboot() returns to SBE Update code for the BMC_IPMI
config and then calls masterVersionCompare().  The masterVersionCompare()
call should not be made if a shutdown has been requested, so should be
put in an else leg for the if statement that determines if sbeDoReboot()
is called.

Change-Id: Ib1b1d7a08b7886e9954edd05284a666e23b37796
RTC: 176755
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/44676
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>