Create unique TI RC for all reconfig loops
Creating a new TI RC (SHUTDOWN_DO_RECONFIG_LOOP = 0x513) that
will be used for all cases where HB is shutting down due to some
kind of hardware event with the expectation that the FSP will
execute a reconfig loop and start the IPL over.

Improved the SRC data for the error log that shows up when this
path is hit to include the potential reasons for the failure.

Change-Id: Ie0a38a99ca1ae1eed6460c6c85b21cc9ddbb8f7e
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/44988
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>