Fix attr names to be more correct and allow multiple customize tags
Previously we were pulling the customize information from hb_temp_defaults
but that isn't true anymore. This commit updates some variable names to
reflect that. Also we noticed that if an attribute had multiple tags
in the hb_customize file only the first tag was being processed. This
commit fixes it so that all tags will be processed.

Change-Id: Ifd8269c3b1621ffd62585e2d9c8a74047c070f4b
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/44980
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>