Default to use the master's side for slave sbe starts in OP
On OP systems (any non-sequential sbe update config), we will
use the same side of the sbe seeprom on the slave as what was
used to boot the master.  This will prevent us from seeing the
same failure booting the slaves that we got booting the master
in the case where we have a bad sbe image for some reason.

resolves zaius-openpower/#61

Change-Id: Ibcd00c43958d0ba2ce5978b0ab921a934f9f58a5
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/45264
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>