Add terminate option for mfg stop on error case
Adding a new config option HANG_ON_MFG_SRC_TERM that will affect
how we handle the MFG_SRC_TERM flag.  The default (and current)
behavior is for Hostboot to put itself into a hang state in order
to prevent a reboot from the BMC.  The new option will allow us
to terminate normally (with a TI) instead.

Change-Id: I82a0062cf50e4161d0da811a823dd7664dc2a5de
CQ: SW401335
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/45793
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>