Implement FAPI_PLAT_CALL_SUBROUTINE macro
HWP writers need a way to request certain HWPs to run on the SBE
for non-master processor targets while we are in secureBoot mode.
This macro will check these conditions and if we are on a non-master
in secureMode then we will issue a fifo chipop to the slave sbe
in order to request the SBE to perform the HWP. This commit also
adds some simple test cases to verify the failure paths. Also
fapi2 test cases are re-enabled in this commit.

Change-Id: Ic42fe2df5692fb5994985d92b2fd767ea7c1234a
RTC: 179062
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/46580
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>