Remove remaining MPIPL workarounds from host_discover_targets
This commit removes some code that was doing workarounds for the
MPIPL path. One workaround was storing ring data in OCC sram, that
is no longer needed. The other workaround was clearing the PM_EXIT
sign on the PM complex's SICR register. This workaround requires
a fix in the hw_ref image which isnt in a driver yet so this
will fail until that is available in fsp-ci

Change-Id: I500d0271f1f61844a96f945c81b4b2d41225051e
RTC: 171340
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/48393
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>