Check parents of OBUS_BRICK to determine their functionality
A situation has occurred where an OBUS was not functional but its
child OBUS_BRICKs were, which doesn't make sense. This fix is to
implement checking for OBUS (parent) functionality as part of
OBUS_BRICK check and mark the OBUS_BRICKs not functional if their
parent is not functional.

Change-Id: I7f52927455f08296e55d73312e932f9ebbd2b700
CQ:SW404097
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/48415
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>