Manually perform multicast scom ops to avoid chip bug
P9 has a bug in the multicast logic that causes it to return a
'chiplet offline' error if there are any chiplets in the multicast
group that are offline.  If the scom is performed via the PIB
we will see a pib error but the data will be valid.  If the scom
is performed via the ADU we will see an error but the data we
get returned will be all FFs.

This commit will convert a multicast read operation into a series
of unicast scoms when the xscom path is in use.

Resolves boston/622

Change-Id: I33cb2cf8c04eb63014b5b67cdc2967ce0fe4c489
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/48434
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>