Re-apply p9_int_scom inits on shutdown path (2nd attempt)
The interrupt resource provider shutdown path is clearing
out some scom inits that need to be set or else we see
problems in the hypervisor. This commit registers a new
event to the shutdown command which will re-apply the
scominits prior to shutting down and passing off control
to the hypervisor.

Change-Id: Ic6bfd1d95b593aee3f6202f5902dcff7adb29b08
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/48495
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>