Add hbHypCommArea struct to end of Hostboot Data reserved mem section
The Hypervisor needs a way to send down an address to the host so that
the host can account for the HRMOR moving and still be able to boot
on an MPIPL. To do this we are adding another subsection to the
existing reserved memory section which already was holding VPD and
attribute info. This subsection will serve as the HB-HYP communication
region. For now PHYP is the only hypervisor using this space and they
are only using it to store the HRMOR if they decide to move it. In order
for the hypervisor to easily find where to write the HRMOR we added an
address pointer to HDAT that hostboot will fill in to tell PHYP where to
write the HRMOR. This is all setting up for us to be able to read the
HRMOR after an MPIPL so that we can load the hypervisor in at the correct
address.

Change-Id: I5d3687a207296bd87d2a7120dd17505e0022b748
RTC: 180959
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/48651
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>