Read the hb/hyp comm area to see if hyp moved hrmor
During MPIPL it is possible that the hrmor was moved while the
hypervisor was running. Before this commit if the hrmor was moved
then hostboot would fail when we tried to collect dump information.
With this commit hostboot will read the hyp/hb comm area to see if
the hypervisor has moved the hrmor and told HB, if the hyp has then
hostboot will set the ATTR_PAYLOAD_BASE attribute on the system
target during the target service initialization

Change-Id: I89b76913ffa106656f188364a2a077f982474084
RTC: 180959
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/48699
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>