Re-enable fapi2 test case module
We determined the problem to be with the getVPD test case when
we tried to get the section info for the MEMD pnor section the
section was not currently loaded. The reason this was sometimes
passing was because the PNOR tests were loading and unloading
the MEMD module. My guess was that we were getting lucking
and reading MEMD while the pnor test has the section loaded

Change-Id: I50c12993d639554ab919ec01bba45e5f4aae761c
RTC: 181401
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/48955
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>