Allow fuzzy matches to WOF tables in Simics
In Simics we deliberately use fewer parts than the module would
normally support for better performance.  This downbin means that
we no longer are using a supported sort when we look for the WOF
tables.  Since we don't really model the voltage/current of the
cores there is no harm in using a different table.

Change-Id: If5b6db6fee1aed6a176a4aca470b1c82c95ee586
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/48972
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>