Clean up things that were left after updating to simics nimbus 2.0
A few workarounds were put in place to get simics moved up to the
2.0 nimbus model. One was that we put in a hack into a mvpd test
case, the other was disabling the occ_xstop handler from starting.
I found that after putting in good mvpd the occ_xstop handler was
booting up okay. An the mvpd test case is fixed to correctly find
the bucketValue from the value returned from the POUNDV_DATA attr

Change-Id: I0eb023c386066c92e475c3302b2f18d984439d3b
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/49071
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>