Modify eSEL script to not ignore cases for command line options
By default Perl's Getopt::Long will ignore case for option names. This
causes conflict between for '-P' and '-p' options that are provide BMC
password and decode option respectively. Below is an example of failed
run with command having both '-P' and '-p' used:

$./eSEL.pl -t <ip-addr>  -p get_and_decode_ami -U ADMIN -P ADMIN
***ERROR: Incorrect option provided.
Specify an option with -p <option>.

To fix this the Getopt::Long import is configured to not ignore the
case of option names by using the 'no_ignore_case' module config
option at import time. With the fix the command above works as
expected:

$./eSEL.pl -t <ip-addr>  -p get_and_decode_ami -U ADMIN -P ADMIN
Found 69 eSELs with PEL data

Resolves #122

Signed-off-by: Vaibhav Jain <vaibhav@linux.vnet.ibm.com>
Change-Id: I6996d3d26d1449c758c50cbbf3720f7351a14edb
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/49543
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>