During MPIPL check if SBE has set Hyp HRMOR in blComm area
It is possible for the Hypervisor to move the HRMOR during runtime
If this occurs, then we trigger an MPIPL, hostboot needs to be
sure to set the PAYLOAD_BASE attribute to whatever the Hypervisor
had the HRMOR set to before mpipling. The HRMOR is read out by
the SBE during its MPIPL steps and written to the bootloader which
the bootloader passes up to hostboot.

Change-Id: I415b4a0465020f89f95fcde974123362067cbbca
RTC:180959
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/49562
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>