No longer pass master ex to HWSV as is no longer needed in p9
In p8 we sent hwsv a message that told them what the master ex for
the proc was. In p9 hwsv needs the master core, and not the master
ex. We already put the change in to set the core, this commit is
just removing the setting of the ex

Change-Id: Ibc1d9cc891ad685b3a3992eeab67e5290d349ccb
RTC: 149318
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/49566
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Brian E. Bakke <bbakke@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>