Remove limit on max size of PNOR TOC
Previously we assumed the TOC would never be larger than 4kb, this
limited us to <31 sections in pnor. We are getting to the point that is
not enough. To satify the github issue a change was needed in the ffs
tool to correctly build the TOC so it describes its size correctly.
Those changes can be found here: github.com/open-power/ffs/pull/14
The changes in this commit fix hostboot so that it will correctly
parse TOCs > 4kb. Resolves open-power/zaius-openpower/#57.

Change-Id: Id2d8c250b0b1e6b87adfd4d927119fe006e8910d
RTC: 180032
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/49926
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>