Zeppelin BUP updates: mss_memdiags workaround
Currently, HWPs are not enabling host attentions, which
causes a long hang in mss_memdiags and do not correctly
mask out the CHIFIR bits. This commit runs the two scoms
to do that before calling mss_memdiags. The real fix will
come in form of a HWP fix via SW405858

Change-Id: I9159e4d757174002b213068c487623c45be9bf75
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/49949
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>