Test and verify concurrent init of PM complex
This testcase tests the path during a concurrent code update, where
PHYP will call HBRT to reload+restart the PM Complex.
The reload path needs to replace all of the code images while preserving
the runtime portions of the data that PHYP has updated.
This testcase verifies that portion of memory was left untouched.

Change-Id: If57c3e2a3fe469028a18ab6829ca6e178ce12940
RTC:168054
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/50214
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Brian E. Bakke <bbakke@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>