TI w/ unrecoverable error if system boots w/ mismatching processors
We do not support processor modules w/ different EC levels to be
on the same system. The rule is that all processors must match
the master processor. If this is not true then hostboot will TI.
Techically we could probably still boot and just gard out the
non-matching proc's but the use case for this will be lab/manufac.
For those cases we want the technician to know right away that they
put a module w/ the wrong DD level in a system.

RTC: 183243
Change-Id: I09b30550edd30c8523fd2a709c7b7a83f597eab8
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/50167
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>