Ignore #V power sorts for op910 only
Some of the parts shipping for op910 have incorrect
power sorts written to #V MVPD -- however they are compatible
as long as the nominal core freq and nest frequency are the same.
Thus remove a FW check to ensure common power sorts across sockets
to allow the parts to work together.

Change-Id: I9dc897d4e751266706437ac7c142f9f253476dc6
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/50471
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>