Set hostboot_mode bit on P3PC interrupt scom reg during intrrp init
The interrupt resource provider is expecting hostboot to set this
bit in the P3PC register to notify the interrupt logic that we
are running hostboot. Setting this allows the interrupt code to
allow thread contexts to be enabled w/o NVT data structures, also
complex store operations to thread context are considered invalid
and are dropped. This bit gets cleared in the shutdown path when
we reset the interrupt logic.

Change-Id: I41430c836633796d6841a965ef634f26939c6eb5
CQ: SW410032
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/50598
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>