Modify how POWERVM is processed in the MCL manager
POWERVM is verified and loaded to its final location before the MCL
processing occurs. Therefore we only need to get the lid sizes and
create HDAT entries
Additionally it was found that sorting the CompInfoMap resulted to
direct key lookup to not work for the POWERVM component, so removed it.

Change-Id: I5008a1eaa01e0910743fdb11fa73b73a26eccd63
RTC: 181900
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/50644
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>