Adding advanced restart logic after detecting vital attentions
This code sets up and calls the SbeRetryHandler class to attempt
an SBE restart. Additionally, we need more than one way to restart
the SBE: p9_start_cbs and hreset. This commit sets up the structure
to use one of the two methods while waiting on hreset to be available.

Change-Id: I17dfb4047268efa84e388391698b668cec2e13e4
RTC:180243
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/50953
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>