Lock i2c around vpd write message to FSP at runtime
Adding this lock mechanism to handle contention between
the Host and OCC. Using the Host's lock as a proxy for any
operations we want the FSP to do for us. HBRT will now call
to lock the master i2c, then call down to the FSP to do the
VPD write, then call to unlock the master.

Change-Id: Ifa06b3bf12fe9914980d39d2663f032ddb950b94
RTC:184131
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/51041
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>