Send FSP relevant information for PSI Diagnotics test when TCEs are enabled
This commit sovles the Hostboot side of Defect SW411967 where the FSP needs
to know the real memory address associated with the PAYLOAD TCE Tokens so
that it can run the PSI Diagnostics test.  The hostboot code will also
temporarily open an unsecure memory region for this test, as well.

Change-Id: Id906e6570beeefbc23a26b2c2421684e8d2f9c2d
Backport:release-fips910
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/51249
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>