xip_customize and TOR API: Improved DD level verification
This update propagates the extracted EC level in xip_customize
through to the TOR API via the call to tor_get_single_ring(). The
TOR API has been updated to verify that the requested DD level
matches matches the DD level in the TOR header of the supplied
ring section which, in the case of xip_customize, is an .overlays
ring section.

Key_Cronus_Test=XIP_REGRESS

Change-Id: Ie5599239d9fb82b84dc3ac0cdbfc7ef4db569496
Original-Change-Id: Ic087c269306acc0afcb8c759fe68eac63ae93d29
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/51339
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Cronus HW CI <cronushw-ci+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>