Corrected issue with hb errlparser not exiting on error
The code was doing a lot of negative one returns and exits when it
should be passing values from 0 - 255 (0 being success and 1 to 255
indicating failure.)  Also, why return a different error when you
have a perfectly good error code already there. So I return the
current error code.

Change-Id: I485e6abcfded39ad8121de109f16e587c53f1f81
CQ: SW326835
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/51507
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Elizabeth K. Liner <eliner@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>