Remove is_master conditional from getVersionInfo in sbe_update
Previously we were having issues reading from the slave proc's
SBE while we were in cache containment. We have re-order the IPL
steps to reset the memory bars later on in the boot. This allows
us to perform the reads now. We had issues running chipOps on simics.
I am opening up a defect to look into this problem but for now we will
only do i2c reads on simics.

Change-Id: I177082d9c443d83ee750c57cde9a16c09844807f
RTC: 182266
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/52040
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>