During MPIPL expand memory size to be FULL_CACHE right away
When attempting MPIPL on Zeppelin we saw that we were running out
of memory during the Hostboot init phase of the MPIPL flow. This
was happening because HB thinks it needs to be running out of
limited cache still at this point. During the attrrp init code
in MPIPL we will page in a bunch of ATTR data. This fragments our
page table and make future large mallocs impossible. This commit
expands Hostboot's memory out to full cache size right away when
hostboot is getting initialized. This will give us some breathing
room during the initialization steps until we hit thread_activate
where we expand the cache out to REAL_MEMORY size

Change-Id: I1fcc33359de989a0ded231b52272b85e06725e18
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/52056
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>